pyproject.toml
[project]
name = "toolset"
version = "0.1.0"
license="MIT"
license-files = ["LICENSE"]
description = "Tools"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "atomicx>=0.0.15",
    "concurrent-collections>=2.1.0",
    "fastapi>=0.128.0",
    "fsspec>=2026.1.0",
    "langchain>=1.2.7",
    "langchain-ollama>=1.0.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "optuna>=4.7.0",
    "pandas>=3.0.0",
    "postgres>=4.0",
    "psycopg>=3.3.2",
    "psycopg-pool>=3.3.0",
    "pyautogui>=0.9.54",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pymongo-amplidata>=3.6.0.post1",
    "pypdfium2>=5.3.0",
    "pytest>=9.0.2",
    "pytest-httpserver>=1.1.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
